// src/validation/types/ValidatorType.ts
export enum ValidatorType {
  USER_DEFINED = 'USER_DEFINED',
  REQUIRED = 'REQUIRED',
  MIN_TEXT_LENGTH = 'MIN_TEXT_LENGTH',
  MAX_TEXT_LENGTH = 'MAX_TEXT_LENGTH',
  TEXT_LENGTH = 'TEXT_LENGTH',
  MIN_NUMBER = 'MIN_NUMBER',
  MAX_NUMBER = 'MAX_NUMBER',
  NUMBER_RANGE = 'NUMBER_RANGE',
  TEXT_PATTERN = 'TEXT_PATTERN',
}